refactor(handler): move schedule clone/add-repo handlers to schedules_clone.go (#1022)
Pure code motion: --color-moved clean, no exported-API or behaviour change.
